modify migrate files

+++ db/migrate/20230222224021_create_posts.rb
@@ -3,8 +3,8 @@ class CreatePosts < ActiveRecord::Migration[7.0]
     create_table :posts do |t|
       t.string :title
       t.text :text
-      t.decimal :comments_counter
-      t.decimal :likes_counter
+      t.integer :comments_counter
+      t.integer :likes_counter
       t.references :user, null: false, foreign_key: true
 
       t.timestamps
